Create 045-AzureStackHCI
This hack is designed to help you get hands-on experience with Azure Stack HCI and all of it's components without the need for Hardware. In this hack you will learn and gain hands on experience with Azure Stack HCI components like Hyper-V, Failover Clustering, Storage Spaces Direct, Software Defined Networking, Azure Kubernetes Services on HCI and more. This hack will continue to evolve and is your go to place to try out things in a lab, before implementing them in real hardware.

The hack starts off by covering the basics of ASHCI and then digs deep into the different components. You will encounter different types of solutions that is or could be needed in a HCI environment.